% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_excel.R
\name{convert_to_excel}
\alias{convert_to_excel}
\title{Convert to Excel}
\usage{
convert_to_excel(pdf_path, output_path)
}
\arguments{
\item{pdf_path}{Path of input PDF}

\item{output_path}{Path to save new Excel file to}
}
\value{
a Excel workbook (as a side effect)
}
\description{
Convert to Excel
}
\keyword{internal}
