% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_powerpoint.R
\name{convert_to_powerpoint}
\alias{convert_to_powerpoint}
\title{Convert to PowerPoint}
\usage{
convert_to_powerpoint(pdf_path, output_path)
}
\arguments{
\item{pdf_path}{Path of input PDF}

\item{output_path}{Path to save new PowerPoint file to}
}
\value{
a print statement (as a side effect)
}
\description{
Convert to PowerPoint
}
\keyword{internal}
