% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_word.R
\name{convert_to_word}
\alias{convert_to_word}
\title{Convert to Word}
\usage{
convert_to_word(pdf_path, output_path)
}
\arguments{
\item{pdf_path}{Path of input PDF}

\item{output_path}{Path to save new Word file to}
}
\value{
print statement (as a side effect)
}
\description{
Convert to Word
}
\keyword{internal}
