% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_kb.R
\name{format_kb}
\alias{format_kb}
\title{Format a file size in KB}
\usage{
format_kb(x)
}
\arguments{
\item{x}{Numeric size in kilobytes.}
}
\value{
A character string, e.g. \code{format_kb(1131.987)} returns \code{"1,132"}.
}
\description{
Rounds to 0 decimal places and adds comma separators.
}
\keyword{internal}
