% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magick_formats.R
\docType{data}
\name{magick_formats}
\alias{magick_formats}
\title{Magick Formats}
\format{
An object of class \code{character} of length 14.
}
\usage{
magick_formats
}
\description{
Character vector of accepted image format extensions
}
\keyword{datasets}
