% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_check.R
\name{package_check}
\alias{package_check}
\title{Package check}
\usage{
package_check(pkg_name, bookmarks = FALSE, silent = FALSE)
}
\arguments{
\item{pkg_name}{Package name in characters}

\item{bookmarks}{Default FALSE. Whether to show shiny notification for bookmarks}

\item{silent}{Default FALSE. Whether to show shiny notification}
}
\value{
a logical.
}
\description{
Package check
}
