% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_placeholder_text.R
\docType{data}
\name{page_placeholder_text}
\alias{page_placeholder_text}
\title{Page Placeholder Text}
\format{
An object of class \code{character} of length 1.
}
\usage{
page_placeholder_text
}
\description{
UI place holder text for the pages input
}
\keyword{datasets}
