% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pages_to_remove.R
\name{parse_pages_to_remove}
\alias{parse_pages_to_remove}
\title{Parse Pages of a PDF}
\usage{
parse_pages_to_remove(input_string)
}
\arguments{
\item{input_string}{String containing pages to remove}
}
\value{
a numeric vector containing pages to remove
}
\description{
Parse Pages of a PDF
}
\keyword{internal}
