% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_images.R
\name{convert_to_images}
\alias{convert_to_images}
\title{Convert to Images}
\usage{
convert_to_images(pdf_path, output_dir, dpi = 300)
}
\arguments{
\item{pdf_path}{Path of input PDF}

\item{output_dir}{Path to save new PowerPoint file to}

\item{dpi}{dots per inch for .png files, default 300}
}
\value{
a path to either a .zip file (if multiple images) or a single .png file
}
\description{
Convert to Images
}
\keyword{internal}
