% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_filename.R
\name{sanitize_filename}
\alias{sanitize_filename}
\title{Sanitize File names}
\usage{
sanitize_filename(filename)
}
\arguments{
\item{filename}{String containing a user-provided file name}
}
\value{
a sanitized character for the file name
}
\description{
Sanitize File names
}
