% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_disk_space.R
\name{sum_disk_space}
\alias{sum_disk_space}
\title{Sums up disk spaces used for PDF files}
\usage{
sum_disk_space(file_list)
}
\arguments{
\item{file_list}{Names list of PDF files}
}
\value{
A character containing total size in kilobytes
}
\description{
Sums up disk spaces used for PDF files
}
