% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark_stamp.R
\name{watermark_stamp}
\alias{watermark_stamp}
\title{Watermark Stamp}
\usage{
watermark_stamp(
  input_pdf,
  output_pdf,
  watermark_text,
  watermark_fontsize = 50,
  watermark_col = "gray80",
  fallback_col = "gray80",
  watermark_alpha = 0.6,
  watermark_rot = 45,
  watermark_fontface = "bold",
  watermark_height = 11,
  watermark_width = 8.5
)
}
\arguments{
\item{input_pdf}{Path of input PDF}

\item{output_pdf}{Path of output PDF}

\item{watermark_text}{Watermark text}

\item{watermark_fontsize}{Watermark fontsize}

\item{watermark_col}{Watermark color}

\item{fallback_col}{Fallback color to use}

\item{watermark_alpha}{Watermark alpha}

\item{watermark_rot}{Watermark rotation angle}

\item{watermark_fontface}{Watermark fontface (one of "plain", "bold", "italic", "bold.italic")}

\item{watermark_height}{Watermark overlay height (inches)}

\item{watermark_width}{Watermark overlay width (inches)}
}
\value{
a path of output PDF
}
\description{
Watermark Stamp
}
